refactor: separate Projects page components into new files
