+++ src/components/ProjectsContent/IProjectsContent.ts
@@ -1,0 +1,7 @@
+import {ComponentProps} from 'react'
+import {Project} from '../../models'
+
+export interface IProjectsContent extends ComponentProps<'article'> {
+    baseProjectsLength: number
+    projects: Project[]
+}
