+++ src/components/ProjectsContent/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectsContent'
