+++ src/components/ProjectsHeader/IProjectsHeader.ts
@@ -1,0 +1,2 @@
+export interface IProjectsHeader extends ComponentProps<T> {
+}
