+++ src/components/ProjectsHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectsHeader'
