+++ src/components/ProjectsList/IProjectsList.ts
@@ -1,0 +1,6 @@
+import {ComponentProps} from 'react'
+import {Project} from '../../models'
+
+export interface IProjectsList extends ComponentProps<'ul'> {
+    projects: Project[]
+}
