+++ src/components/ProjectsList/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectsList'
