+++ src/components/ProjectsPanel/IProjectsPanel.ts
@@ -1,0 +1,5 @@
+import {ComponentProps} from 'react'
+import {Project} from '../../models'
+export interface IProjectsPanel extends ComponentProps<'dialog'> {
+    project?: Project
+}
